requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.21
passlib[bcrypt]==1.7.4
python-jose==3.3.0
numpy>=1.26.0
Pillow==10.1.0
tqdm==4.66.1
rasterio
folium